* adding new program... punetwiin.  its purpose is kinda like betwiin.  but it works slightly different.  for one, it takes a full nand.bin for both done files.  it pulls the files and filesystem from the first nand, and it pull boot1&2, the keys and the bad block map from the second nand.  then it bakes on low heat for 45 minutes and when its done, you will have the contents from the filesystem in teh first nand mixed with everything else from the second nand.
i only performed limited testing, but it seems to work fine.  in linux, the full process took < 20 seconds.  in windows, it was about 2 minutes.  bad blocks are taken into account, so i personally feel that it is safer that betwiin ( but im not knocking that program at all :) )

anybody using punetwiin is encouraged to verify the content of the output file before even thinking about installing it.

git-svn-id: http://wiiqt.googlecode.com/svn/trunk@83 389f4c8b-5dfe-645f-db0e-df882bc27289